Fix SqlValue ambiguity on Linux x86_64 (long long != long); verify 189/189 tests in Docker (2026-05-09); add reproducer + build log

+++ cobol-runtime/src/sql_context.cpp
@@ -79,7 +79,7 @@ void SqlContext::execute_on_db(const std::string& sql, const std::vector<SqlValu
                         row.values.emplace_back();
                         break;
                     case SQLITE_INTEGER:
-                        row.values.emplace_back(SqlValue(sqlite3_column_int64(stmt, i)));
+                        row.values.emplace_back(SqlValue(static_cast<int64_t>(sqlite3_column_int64(stmt, i))));
                         break;
                     case SQLITE_FLOAT:
                         row.values.emplace_back(SqlValue(sqlite3_column_double(stmt, i)));
